+++ src/withMapper/frequentSetsMapper.java
@@ -1,3 +1,4 @@
+package withMapper;
 import java.io.BufferedReader;
 import java.io.IOException;
 import java.io.InputStreamReader;
