+++ src/withMapper/itemReducer.java
@@ -1,3 +1,4 @@
+package withMapper;
 import java.io.IOException;
 
 import org.apache.hadoop.io.IntWritable;
